Update Broken Links & Add Ruby Tutorial
Hi,
I updated the links to the Head First Java book, which was 404, as well as the link to the official Github cheatsheet, which was also 404. Both links are now working.

I also added a new Ruby tutorial which will be helpful for new Ruby developers looking to learn the language.

Hope that helps!